MR-6: Simplify handling of authentication errors
According to the OAuth spec, the "error" attribute is required, and the "error_description" is optional. Also, neither values are intended to be human readable so are not appropriate to display on the UI. Change to just set a boolean flag for when an error has occurred, and display a fixed error message to the user.

RB/MT