[PRMT-2924]: Add prompt parameter when redirecting to OIDC authorize urls
Includes refactoring the application to use the oidc-client-ts library for OIDC authentication instead of Amplify, because Amplify does not support the prompt parameter. This required us to use Cognito as a genuine OIDC provider in its own right instead of using the Amplify UI components for authenticating with the user pool.

We also took the opportunity to gitignore the config.js file and create an example to copy from instead