[TECH] Enable running of integration tests in pipeline
This required a few substantial changes:
- Addition of run integration test stage in dev pipeline
- Addition of suitable AWS IAM role for running integration tests
- Ability to run AWS CLI and gradle in the same container in the pipeline so that we can assume our new IAM role and run the tests. Hence using a new image containing all the necessary tools for our development.

This permitted several useful simplifications. We no longer needed separate OpenJDK and infra images, we can just use the base dojo file for everything except nodeJS. Therefore we can remove the bootstrap-terraform and terraform service from the docker-compose file. That means storing our terraform state as a file in the local environment, which we do in the Dojo user's home directory, to make sure the state is not retained when the container shuts down.

Co-authored-by: Richard Smith <richard.smith33@nhs.net>